refactor(print-format-builder): extract serialize_layout into utils
